added street_address_only data type and option to set select dropdowns
The available options for values from the geocode request do not include a street only address, e.g., 123 Main St instead of 123 Main St, Anytown, USA. This option was added.

The setDetail function failed to set select dropdowns with the proper value. I personally was using them for state selection, but it should work for anything where the returned value from the geocode is a match for the select option's value or text. This has been added now.