Removed street_address_only address component and added multiple address component support
Added the ability to set multiple address component types to a single input field. The components will be set in the input in the order in which the markup lists them. For example, on a return from google of "123 Main St, Anytown, AA", using the detailsAttribute of `data-geo` and setting `data-geo="street_number route". Behavior remains the same for inputs that have a repeated component type as the first data-geo attribute. Only the first input will be filled.