Fix Google Places API URL in README
From Google Places documentation: The sensor parameter is no longer
required for the Google Maps JavaScript API. It won't prevent the Google
Maps JavaScript API from working correctly, but we recommend that you
remove the sensor parameter from the script element.

Leaving this parameter in displays a Warning in the browser's console.

Source:
https://developers.google.com/maps/documentation/javascript/error-messages#sensor-not-required